#4 refactor:accout api 분리
